Keystone liberty-1 milestone (8.0.0.0b1)

Conflicts:
	keystone/locale/en_AU/LC_MESSAGES/keystone.po
	keystone/locale/pt_BR/LC_MESSAGES/keystone.po

+++ keystone/token/providers/fernet/utils.py
@@ -59,8 +59,8 @@ def _convert_to_integers(id_value):
     try:
         id_int = int(id_value)
     except ValueError as e:
-        msg = ('Unable to convert Keystone user or group ID. Error: %s', e)
-        LOG.error(msg)
+        msg = _LE('Unable to convert Keystone user or group ID. Error: %s')
+        LOG.error(msg, e)
         raise
 
     return id_int
@@ -178,7 +178,7 @@ def rotate_keys(keystone_user_id=None, keystone_group_id=None):
 
     LOG.info(_LI('Starting key rotation with %(count)s key files: %(list)s'), {
         'count': len(key_files),
-        'list': key_files.values()})
+        'list': list(key_files.values())})
 
     # determine the number of the new primary key
     current_primary_key = max(key_files.keys())
@@ -208,11 +208,13 @@ def rotate_keys(keystone_user_id=None, keystone_group_id=None):
 
     # purge excess keys
     keys = sorted(key_files.keys())
-    excess_keys = (
-        keys[:len(key_files) - CONF.fernet_tokens.max_active_keys + 1])
-    LOG.info(_LI('Excess keys to purge: %s'), excess_keys)
-    for i in excess_keys:
-        os.remove(key_files[i])
+    number_of_keys_to_purge = max(
+        0, len(key_files) - CONF.fernet_tokens.max_active_keys + 1)
+    if number_of_keys_to_purge > 0:
+        excess_keys = keys[:number_of_keys_to_purge]
+        LOG.info(_LI('Excess keys to purge: %s'), excess_keys)
+        for i in excess_keys:
+            os.remove(key_files[i])
 
 
 def load_keys():
